requirements.txt
faster-whisper==1.0.0
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio
fastapi
uvicorn[standard]
pydantic
numpy==1.26.4
python-multipart
